+++ server/src/formatting/types.ts
@@ -28,12 +28,6 @@ export type FormattingSettings = {
 	wordWrapColumn: number;
 };
 
-export type FormattingFlags = {
-	removeMacroMultiline: boolean;
-	runLongLineCheck: boolean;
-	runExpressionIndentationCheck: boolean;
-};
-
 export type CustomDocumentFormattingParams = (
 	| DocumentFormattingParams
 	| DocumentRangeFormattingParams