+++ CHANGELOG.md
@@ -6,11 +6,23 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.8.5] - Unreleased
 
+### Added
+
+- Formatting: empty reference node are removed when formatting example
+
+```devicetree
+&n {}; // removed
+l1: &n {}; // kept as this adds label to context
+&n { prop; }; // kept as this adds property to context
+&n { node {}; }; // kept as this node property to context
+```
+
 ### Fixed
 
 - Auto Completion for `compatible` property when using Zephyr bindings.
 - Default value for `status` is ignored when using zephyr bindings.
   The defaults are already defined in the devicetree specification section 2.3.4.
+- Fixed missing formatting diagnostic message when node address has `ull`
 
 ## [0.8.4] - 2026-03-22